separate evolution website homepage + add animation + fix bug on chip and button styles
