+++ apps/website/src/lib/styles/animations/index.ts
@@ -1,0 +1,1 @@
+export * from './scroll';
