+++ packages/lapikit/CHANGELOG.md
@@ -19,7 +19,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.1]
+## [0.2.1] - 2025-07-31
 
 ### Added
 
@@ -29,6 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix **Alert** component bad position content and style not working with bad className.
 - Fix **Chip** component bad content alignement with prepend and append snippet.
+- Fix **Chip** component on outline style background transparent...
+- Fix **Button** component on outline style background transparent...
 
 ## [0.2.0] - 2025-07-26
